batteri

+++ src/globals.cpp
@@ -9,7 +9,7 @@ Zumo32U4Motors motors;
 Zumo32U4LineSensors lineSensors;
 unsigned int sensorValues[5];
 Zumo32U4LCD lcd;
-int battery_cap=100;
+float battery_cap=100;
 const float power=0.0005; // kan justeres for hvor fort batteriet skal tømmes
 uint16_t threshold = 700;
 int baseSpeed = 100;
